move AllTransaction struct and relate types into a single file
