+++ crates/transaction-pool/src/pool/blob.rs
@@ -1,4 +1,4 @@
-use super::txpool::PendingFees;
+use super::all_tx::PendingFees;
 use crate::{
     identifier::TransactionId, pool::size::SizeTracker, traits::BestTransactionsAttributes,
     PoolTransaction, SubPoolLimit, ValidPoolTransaction,
